requirements.txt
numpy==1.26.1
tqdm==4.67.1
opencv-python==4.11.0.86
cffi==1.16.0
defusedxml==0.7.1
pillow==10.3.0
ipython==8.24.0
jupyter==1.0.0
pip==25.0.1
attrs==23.2.0
distro==1.9.0
wheel==0.43.0
filelock==3.15.4
requests==2.31.0
rich==13.7.1
keras==3.3.3
colorama==0.4.6
ipywidgets==8.1.2
notebook==7.2.0
PyYAML==6.0.1
pytest==8.3.5
Cython==3.0.12
setuptools==69.5.1
pytz==2024.1
overrides==7.7.0
numba==0.61.0
mypy==1.15.0
scipy==1.11.4
threadpoolctl==3.6.0
torch==2.5.1
matplotlib==3.7.5
mpmath==1.3.0
sympy==1.13.1
platformdirs==4.2.2
joblib==1.3.2
psutil==5.9.8
cloudpickle==3.1.1
six==1.16.0
pynput==1.8.1
imageio==2.37.0
tifffile==2025.3.30
scikit-image==0.25.2
networkx==3.4.2
scikit-learn==1.4.2
packaging==24.2
deprecation==2.1.0
pandas==2.1.4
pymsgbox==1.0.9
pyperclip==1.9.0
MouseInfo==0.1.3
pyscreeze==1.0.1
pytweening==1.2.0
PyGetWindow==0.0.9
pyautogui==0.9.54
PyRect==0.2.0
mss==10.0.0